+++ packages/iov-bns/src/bnsconnection.spec.ts
@@ -26,7 +26,7 @@ import {
 } from "@iov/bcp-types";
 import { Random, Sha256 } from "@iov/crypto";
 import { Encoding, Uint64 } from "@iov/encoding";
-import { Ed25519HdWallet, HdPaths, LocalIdentity, UserProfile, WalletId } from "@iov/keycontrol";
+import { Ed25519HdWallet, HdPaths, UserProfile, WalletId } from "@iov/keycontrol";
 import { asArray, lastValue } from "@iov/stream";
 
 import { bnsCodec } from "./bnscodec";
@@ -89,7 +89,7 @@ describe("BnsConnection", () => {
   ): Promise<{
     readonly profile: UserProfile;
     readonly mainWalletId: WalletId;
-    readonly faucet: LocalIdentity;
+    readonly faucet: PublicIdentity;
   }> {
     const wallet = Ed25519HdWallet.fromMnemonic(defaultMnemonic);
     const profile = new UserProfile();