+++ packages/iov-keycontrol/types/index.d.ts
@@ -1,5 +1,5 @@
 export { HdPaths } from "./hdpaths";
 export { Keyring } from "./keyring";
 export { UserProfile, WalletInfo } from "./userprofile";
-export { LocalIdentity, LocalIdentityId, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString, } from "./wallet";
+export { Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "./wallet";
 export { Ed25519HdWallet, Ed25519Wallet, Secp256k1HdWallet } from "./wallets";